Calculate average allele_freq from Dirichlet distribution
combineAF now uses the scipy dirichlet distribution means as the combined allele frequency.
To do this it calculates c for each row as reported allele freq
multiplied by the supplied weights argument, which defaults to
2n (which is also created as sample_size *2).
The default prior is [1]*k.

This function requires a single loci so this is asserted.

The demo files have been updated accordingly.
Mostly filtering to single loci, and removing unmeasured_alleles calls
because these are handled by combineAF().